requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.21.4
botocore==1.24.5
certifi==2022.6.15
Django==3.2.6
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
drf-base64==2.0
httplib2==0.20.4
jmespath==0.10.0
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8